third_party/blink/web_tests/platform/generic/fast/text-autosizing/text-removal-expected.txt

PASS window.getComputedStyle(span1).lineHeight is "16px"
PASS window.getComputedStyle(span2).lineHeight is "16px"
PASS successfullyParsed is true

TEST COMPLETE
